+++ components/salesforce_rest_api/actions/salesforce-sosl-search/salesforce-sosl-search.mjs
@@ -1,28 +1,25 @@
 import salesforce from "../../salesforce_rest_api.app.mjs";
-import { axios } from "@pipedream/platform";
 
 export default {
   key: "salesforce_rest_api-salesforce-sosl-search",
   name: "SOSL Search",
-  description: "Executes the specified SOSL search.",
+  description: "Executes the specified SOSL search. See [docs](https://developer.salesforce.com/docs/atlas.en-us.soql_sosl.meta/soql_sosl/sforce_api_calls_sosl.htm)",
   version: "0.2.2",
   type: "action",
   props: {
     salesforce,
-    sosl_search_string: {
+    search: {
       type: "string",
-      label: "sosl_search_string",
-      description: "The search string in SOSL.\nFor more information on [SOSL see the SOQL and SOSL Reference.](http://www.salesforce.com/us/developer/docs/soql_sosl/index_Left.htm)\nExample: [Search for a String](https://developer.salesforce.com/docs/atlas.en-us.api_rest.meta/api_rest/dome_search.htm).",
+      label: "SOSL Query",
+      description: "A SOSL search query",
     },
   },
   async run({ $ }) {
-    //See the API docs here: https://developer.salesforce.com/docs/atlas.en-us.api_rest.meta/api_rest/intro_what_is_rest_api.htm
-    return await axios($, {
-      url: `${this.salesforce.$auth.instance_url}/services/data/v20.0/search/?q=${encodeURI(this.sosl_search_string)}`,
-      headers: {
-        Authorization: `Bearer ${this.salesforce.$auth.oauth_access_token}`,
-      },
-
+    const response = await this.salesforce.search({
+      $,
+      search: this.search,
     });
+    $.export("$summary", "Returned results for SOSL search");
+    return response;
   },
 };
