+++ components/salesforce_rest_api/common/sobjects/note.mjs
@@ -1,0 +1,17 @@
+export default {
+  Body: {
+    type: "string",
+    label: "Body",
+    description: "Body of the note. Limited to 32 KB.",
+  },
+  IsPrivate: {
+    type: "boolean",
+    label: "IsPrivate",
+    description: "If true, only the note owner or a user with the Modify All Data permission can view the note or query it via the API. Note that if a user who does not have the Modify All Data permission sets this field to true on a note that they do not own, then they can no longer query, delete, or update the note.",
+  },
+  OwnerId: {
+    type: "string",
+    label: "OwnerId",
+    description: "ID of the user who owns the note.",
+  },
+};
