+++ components/salesforce_rest_api/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@pipedream/salesforce_rest_api",
-    "version": "0.3.3",
+    "version": "0.4.0",
     "description": "Pipedream Salesforce_rest_api Components",
     "main": "salesforce_rest_api.app.mjs",
     "keywords": [
@@ -12,7 +12,8 @@
     "license": "MIT",
     "dependencies": {
         "@pipedream/platform": "^0.9.0",
-        "salesforce-webhooks": "^1.1.4"
+        "salesforce-webhooks": "^1.1.4",
+        "lodash": "^4.17.21"
     },
     "gitHead": "e12480b94cc03bed4808ebc6b13e7fdb3a1ba535",
     "publishConfig": {
